+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- [Banner-Image for Group](https://github.com/5pm-HDH/churchtools-api/issues/65)
 
 ### Changed